1 - First App/requirements.txt
langchain == 0.3.3
langchain-openai == 0.2.2
langchain-ollama == 0.2.0
python-dotenv == 1.0.1